TECH-231 / TECH-315 - fixes for modules patch (#704)
* Keep modules patches
* execute it only in case of migration
* Clean up patches filter
* Removed duplicate patch

+++ war/src/data/resources/patches/groovy/0.0.0.0-01-modules.beforeContextInitializing.groovy
@@ -1,9 +1,17 @@
 import org.apache.commons.io.FileUtils
+import org.jahia.bin.Jahia
+import org.jahia.osgi.FrameworkService
 import org.jahia.settings.SettingsBean
+import org.jahia.tools.patches.Patcher
 
-def file = new File(SettingsBean.getInstance().getJahiaVarDiskPath() + "/bundles-deployed")
-if (file.exists()) {
-    FileUtils.deleteQuietly(file)
-    FileUtils.touch(new File(SettingsBean.getInstance().getJahiaVarDiskPath(), "[persisted-bundles].dorestore"))
-    FileUtils.touch(new File(SettingsBean.getInstance().getJahiaVarDiskPath(), "[generated-resources].dodelete"))
+setResult("keep")
+
+if (!FrameworkService.getInstance().isStarted() && !Patcher.getInstance().getJahiaPreviousVersion().equals(Jahia.JAHIA_VERSION)) {
+    def file = new File(SettingsBean.getInstance().getJahiaVarDiskPath() + "/bundles-deployed")
+    if (file.exists()) {
+        log.info("*** Migration detected, cleaning all previous bundles and instructing for modules reinstallation ***")
+        FileUtils.deleteQuietly(file)
+        FileUtils.touch(new File(SettingsBean.getInstance().getJahiaVarDiskPath(), "[persisted-bundles].dorestore"))
+        FileUtils.touch(new File(SettingsBean.getInstance().getJahiaVarDiskPath(), "[generated-resources].dodelete"))
+    }
 }
